backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
redis==5.2.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic[email]==2.10.3
alembic==1.14.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
